size_t srclen to avoid chocking on direct strlen() argument

+++ src/nc_string.h
@@ -46,7 +46,7 @@ void string_init(struct string *str);
 void string_deinit(struct string *str);
 bool string_empty(const struct string *str);
 rstatus_t string_duplicate(struct string *dst, const struct string *src);
-rstatus_t string_copy(struct string *dst, const uint8_t *src, uint32_t srclen);
+rstatus_t string_copy(struct string *dst, const uint8_t *src, size_t srclen);
 int string_compare(const struct string *s1, const struct string *s2);
 
 /*
